+++ packages/editor/src/components/visual-editor/index.js
@@ -10,7 +10,6 @@ import {
 	BlockList,
 	store as blockEditorStore,
 	__unstableUseTypewriter as useTypewriter,
-	__unstableUseTypingObserver as useTypingObserver,
 	useSettings,
 	RecursionProvider,
 	privateApis as blockEditorPrivateApis,
@@ -303,7 +302,6 @@ function VisualEditor( {
 		blockListLayout?.type === 'default' && ! hasPostContentAtRootLevel
 			? fallbackLayout
 			: blockListLayout;
-	const observeTypingRef = useTypingObserver();
 	const titleRef = useRef();
 	useEffect( () => {
 		if ( ! autoFocus || ! isCleanNewPost() ) {
@@ -452,7 +450,6 @@ function VisualEditor( {
 								}
 							) }
 							contentEditable={ false }
-							ref={ observeTypingRef }
 							style={ {
 								// This is using inline styles so it's applied
 								// within the editor iframe.